Changed scale factors from int to qreal
Made tiny=0.5 and small = 0.75

+++ DMHelper/src/battledialogmodelcharacter.cpp
@@ -43,7 +43,7 @@ BattleDialogModelCombatant* BattleDialogModelCharacter::clone() const
     return new BattleDialogModelCharacter(*this);
 }
 
-int BattleDialogModelCharacter::getSizeFactor() const
+qreal BattleDialogModelCharacter::getSizeFactor() const
 {
     Character* character = getCharacter();
     if(!character)
