+++ DMHelper/src/battledialogmodelcharacter.h
@@ -25,7 +25,7 @@ public:
     // Local
     virtual int getType() const override;
     virtual BattleDialogModelCombatant* clone() const override;
-    virtual int getSizeFactor() const override;
+    virtual qreal getSizeFactor() const override;
     virtual int getSizeCategory() const override;
 
     virtual int getStrength() const override;
